+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignite-ui-kevinsilvaa/tokens
 
+## 1.2.0
+
+### Minor Changes
+
+- Add Toast and Tooltip components and a red color for Toast error variant
+
 ## 1.1.0
 
 ### Minor Changes